+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [//]: # "## [unreleased] - yyyy-mm-dd"
 
 
-## [unreleased] - yyyy-mm-dd
+## [0.4.0] - 2025-11-16
 
 ### Added
 - Added support for configurable handling of missing RCF wavelengths by introducing the `MissingRCFBehavior`